// src/api/models/ticketGrantNoticeResDTO.ts
/**
 * Generated by orval v8.4.2 🍺
 * Do not edit manually.
 * Folioo API docs
 * API document of Folioo Development
 * OpenAPI spec version: 1.0.0
 */
import type { TicketGrantNoticePayloadResDTO } from './ticketGrantNoticePayloadResDTO';
import type { TicketGrantNoticeResDTOCtaLink } from './ticketGrantNoticeResDTOCtaLink';
import type { TicketGrantNoticeResDTOCtaText } from './ticketGrantNoticeResDTOCtaText';
import type { TicketGrantNoticeResDTOStatus } from './ticketGrantNoticeResDTOStatus';

export interface TicketGrantNoticeResDTO {
  id: number;
  ticketGrantId: number;
  status: TicketGrantNoticeResDTOStatus;
  title: string;
  body: string;
  /** @nullable */
  ctaText?: TicketGrantNoticeResDTOCtaText;
  /** @nullable */
  ctaLink?: TicketGrantNoticeResDTOCtaLink;
  payload?: TicketGrantNoticePayloadResDTO | null;
  createdAt: string;
}